requirements.txt
streamlit
pillow
qrcode
gspread
oauth2client
requests
